fix: made package.json test more stable by moving it to a separate folder and removed auto generation
