+++ tests/i18n-cli.spec.ts
@@ -376,104 +376,3 @@ describe('version command test', () => {
   });
 });
 
-describe('config file', () => {
-  const generateTypesCommand = new GenerateTypesCommand();
-  let typesOutputPath: string;
-  let mockExit;
-
-  beforeEach(() => {
-    mockExit = jest.spyOn(process, 'exit').mockImplementation(() => {
-      return undefined as never;
-    });
-
-    const randomNumber = Math.floor(Math.random() * 1000000000);
-    typesOutputPath = path.join(
-      os.tmpdir(),
-      `/generated/i18n-${randomNumber}.generated.ts`,
-    );
-  });
-
-  afterEach(() => {
-    try {
-      fs.unlinkSync(typesOutputPath);
-    } catch (e) {}
-
-    jest.clearAllMocks();
-  });
-
-  it('should use options file', async () => {
-    await generateTypesCommand.handler({
-      typesOutputPath: typesOutputPath,
-      watch: false,
-      debounce: 200,
-      translationsPath: [],
-      loaderType: [],
-      optionsFile: path.join(__dirname, './app/config/i18n-options.ts'),
-    } as yargs.Arguments<GenerateTypesArguments>);
-
-    const newFileContent = fs.readFileSync(typesOutputPath).toString();
-
-    expect(newFileContent).toContain(`"validation": {
-        "email": string;
-        "password": string;
-        "NOT_EMPTY": string;
-        "INVALID_EMAIL": string;
-        "INVALID_BOOLEAN": string;
-        "MIN": string;
-        "MAX": string;
-    };`);
-  });
-
-  it('should use nearest package.json file', async () => {
-    try {
-      fs.unlinkSync(`/tmp/i18n-generated.ts`);
-    } catch (e) {}
-
-    const mockCwd = jest.spyOn(process, 'cwd').mockImplementation(() => {
-      return __dirname;
-    });
-
-    const packageJsonPath = path.join(__dirname, `./package.json`);
-    const outputPath = `/tmp/i18n-${Math.floor(Math.random() * 1000000000)}-generated.ts`;
-
-    fs.writeFileSync(
-      packageJsonPath,
-      JSON.stringify({
-        i18n: {
-          typesOutputPath: outputPath,
-          optionsFile: './app/config/i18n-options.ts',
-        },
-      }),
-    );
-
-    while (true) {
-      const buffer = fs.readFileSync(packageJsonPath);
-      if(buffer.length > 0) break;
-      await new Promise((resolve) => setTimeout(resolve, 100));
-    }
-
-    await generateTypesCommand.handler({
-      typesOutputPath: '',
-      watch: false,
-      debounce: 200,
-      translationsPath: [],
-      loaderType: [],
-    } as yargs.Arguments<GenerateTypesArguments>);
-
-    const newFileContent = fs.readFileSync(outputPath).toString();
-
-    expect(newFileContent).toContain(`"validation": {
-        "email": string;
-        "password": string;
-        "NOT_EMPTY": string;
-        "INVALID_EMAIL": string;
-        "INVALID_BOOLEAN": string;
-        "MIN": string;
-        "MAX": string;
-    };`);
-
-    try {
-      fs.unlinkSync(packageJsonPath);
-    } catch (e) {}
-  });
-});
